+++ core/cross-client/src/task.rs
@@ -1,10 +1,6 @@
-use std::sync::Arc;
-
 use reqwest::Url;
 
-use protocol::types::{ValidatorExtend, H160};
-
-use crate::types::Requests;
+use protocol::types::{ValidatorExtend, H160, Requests};
 
 pub struct RequestCkbTask {
     address:     H160,
