+++ core/storage/src/adapter/rocks.rs
@@ -294,6 +294,7 @@ const C_RECEIPTS: &str = "c4";
 const C_WALS: &str = "c5";
 const C_HASH_HEIGHT_MAP: &str = "c6";
 const C_EVM_CODE_MAP: &str = "c7";
+const C_CROSSCHAIN_RECORD_MAP: &str = "c8";
 
 fn map_category(c: StorageCategory) -> &'static str {
     match c {
@@ -304,6 +305,7 @@ fn map_category(c: StorageCategory) -> &'static str {
         StorageCategory::Wal => C_WALS,
         StorageCategory::HashHeight => C_HASH_HEIGHT_MAP,
         StorageCategory::Code => C_EVM_CODE_MAP,
+        StorageCategory::Crosschain => C_CROSSCHAIN_RECORD_MAP,
     }
 }
